clearing_house: signed fee pool (#27)
* init

* add max spread

* fix bugs

* max spread uses marginRatioInitial

* more max spread improvements/amm.maxSpread

* fix extra arg

* add checked math

* update mark twap with optimistic bid/ask

* more mark twap tests

* expand test for twap

* update_pool_balances

* fix settle bug

* make max spread logic consistent to amm.max_spread

* more test invariants

* fix test

* clean up

* clean code/logic

* improve naming

* add repeg.rs

* add test to run-anchor-tests, improve repeg cost rounding

* update quote bank reserve in settle_pnl

* remove dead comment

* remove redudant

+++ programs/clearing_house/src/controller/orders.rs
@@ -812,7 +812,7 @@ pub fn fulfill_order_with_amm(
     market.amm.total_fee_minus_distributions = market
         .amm
         .total_fee_minus_distributions
-        .checked_add(fee_to_market)
+        .checked_add(fee_to_market as i128)
         .ok_or_else(math_error!())?;
     market.amm.net_revenue_since_last_funding = market
         .amm
@@ -995,7 +995,7 @@ pub fn fulfill_order_with_match(
     market.amm.total_fee_minus_distributions = market
         .amm
         .total_fee_minus_distributions
-        .checked_add(fee_to_market)
+        .checked_add(fee_to_market as i128)
         .ok_or_else(math_error!())?;
     market.amm.net_revenue_since_last_funding = market
         .amm
